// frontend/src/client/types.gen.ts
export type GiftCreate = {
  name: string
  description?: string | null
  approximate_price?: number | null
  photo_url?: string | null
  product_link?: string | null
}